[server] Merge the ak challenge and the quote nonce
Merges the secret generated for the credential activation challenge and
the anti-replay nonce in the quote.

The client stills generate the MakeCredential challenge, and still sends
it to the server, but now, rather than responding with the decrypted
secret, the server immediately generates the quote, with the decrypted
challenge as the anti-replay nonce, and sends this back to the client.
The client now has to check that the nonce contained in the quote is the
same that the one generated by the MakeCredential call.

This makes the protocol ligher, removing two unneeded messages.

Signed-off-by: lfalkau <lfalkau@student.42.fr>